Do deploy from samples
