+++ Samples/Mapsui.Samples.Common/Maps/Data/MbTilesOverlaySample.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using BruTile.MbTiles;
 using BruTile.Predefined;
+using Mapsui.Samples.Common.Utilities;
 using Mapsui.Tiling.Layers;
 using Mapsui.UI;
 using SQLite;
@@ -10,6 +11,11 @@ namespace Mapsui.Samples.Common.Maps
 {
     public class MbTilesOverlaySample : ISample
     {
+        static MbTilesOverlaySample()
+        {
+            MbTilesDeployer.CopyEmbeddedResourceToFile("torrejon-de-ardoz.mbtiles");
+        }
+
         public string Name => "2 MbTiles Overlay";
         public string Category => "Data";
 
@@ -17,9 +23,10 @@ namespace Mapsui.Samples.Common.Maps
         {
             var map = new Map();
             map.Layers.Add(new TileLayer(KnownTileSources.Create(KnownTileSource.BingAerial, persistentCache: BingArial.DefaultCache)) { Name = "Bing Aerial" });
-            map.Layers.Add(CreateMbTilesLayer(Path.Combine(MbTilesSample.MbTilesLocation, "torrejon-de-ardoz.mbtiles")));
+            map.Layers.Add(CreateMbTilesLayer(Path.Combine(MbTilesDeployer.MbTilesLocation, "torrejon-de-ardoz.mbtiles")));
             return Task.FromResult(map);
         }
+
         public static TileLayer CreateMbTilesLayer(string path)
         {
             var mbTilesTileSource = new MbTilesTileSource(new SQLiteConnectionString(path, true));
