+++ Samples/Mapsui.Samples.Common/Mapsui.Samples.Common.csproj
@@ -83,6 +83,10 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Helpers\" />
+  </ItemGroup>
+
   <PropertyGroup>
     <DebugType>portable</DebugType>
   </PropertyGroup>
