+++ Samples/Mapsui.Samples.Forms/Mapsui.Samples.Forms.WPF/MainWindow.xaml.cs
@@ -1,5 +1,5 @@
-﻿using Mapsui.Samples.Common.Helpers;
-using Mapsui.Samples.Common.Maps;
+﻿using Mapsui.Samples.Common.Maps;
+using Mapsui.Samples.Common.Utilities;
 using System;
 using System.IO;
 using Xamarin.Forms.Platform.WPF;
@@ -17,13 +17,7 @@ namespace Mapsui.Samples.Forms.WPF
 
             Xamarin.Forms.Forms.Init();
 
-            // Hack to tell the platform independent samples where the files can be found on iOS.
-            MbTilesSample.MbTilesLocation = MbTilesLocationOnWPF;
-            MbTilesHelper.DeployMbTilesFile(MbTilesLocationOnWPF);
-
-            LoadApplication(new Mapsui.Samples.Forms.App());
+            LoadApplication(new Forms.App());
         }
-
-        private static string MbTilesLocationOnWPF => Environment.GetFolderPath(Environment.SpecialFolder.Personal);
     }
 }
