+++ Samples/Mapsui.Samples.Forms/Mapsui.Samples.Forms.iOS/AppDelegate.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.IO;
 using Foundation;
-using Mapsui.Samples.Common.Helpers;
 using Mapsui.Samples.Common.Maps;
+using Mapsui.Samples.Common.Utilities;
 using UIKit;
 
 namespace Mapsui.Samples.Forms.iOS
@@ -24,15 +24,9 @@ namespace Mapsui.Samples.Forms.iOS
         {
             global::Xamarin.Forms.Forms.Init();
 
-            // Hack to tell the platform independent samples where the files can be found on iOS.
-            MbTilesSample.MbTilesLocation = MbTilesLocationOnIos;
-            MbTilesHelper.DeployMbTilesFile(MbTilesLocationOnIos);
-
             LoadApplication(new App());
 
             return base.FinishedLaunching(app, options);
         }
-
-        private static string MbTilesLocationOnIos => Environment.GetFolderPath(Environment.SpecialFolder.Personal);
     }
 }
